% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_datastore.r
\name{authenticate_datastore}
\alias{authenticate_datastore}
\title{Authenticate Datastore}
\usage{
authenticate_datastore(key, secret, project)
}
\arguments{
\item{key}{OAuth 2.0 credential key}

\item{secret}{OAuth credential secret key}

\item{project}{Google cloud platform project id/name.}
}
\description{
Authenticate datastore using OAuth 2.0. Create an application on the
\strong{google cloud platform}
and generate
}
\seealso{
\url{https://cloud.google.com/}

\url{https://cloud.google.com/datastore/docs/concepts/overview}

\url{https://developers.google.com/identity/protocols/OAuth2#basicsteps}
}

