% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_aws_s3.r
\name{cache_aws_s3}
\alias{cache_aws_s3}
\title{Amazon Web Services S3 Cache}
\usage{
cache_aws_s3(cache_name)
}
\arguments{
\item{cache_name}{Bucket name for storing cache files.}
}
\description{
Initiate an Amazon Web Services Cache
}
\examples{

\dontrun{
# Set AWS credentials.
Sys.setenv("AWS_ACCESS_KEY_ID" = "<access key>",
           "AWS_SECRET_ACCESS_KEY" = "<access secret>")

# Set up a unique bucket name.
s3 <- cache_aws_s3("unique-bucket-name")
mem_runif <- memoise(runif, cache = s3)
}


}

