% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_postgresql.R
\name{cache_postgresql}
\alias{cache_postgresql}
\title{PostgreSQL Cache
PostgreSQL-backed cache, for remote caching.}
\usage{
cache_postgresql(pg_con, table_name, algo = "sha512", compress = FALSE)
}
\arguments{
\item{pg_con}{An RPostgreSQL connection object}

\item{table_name}{A postgres table name with val and key fields. May include a schema.}

\item{algo}{The hashing algorithm used for the cache, see
\code{\link[digest]{digest}} for available algorithms.}

\item{compress}{Argument passed to \code{saveRDS}. One of FALSE, "gzip",
"bzip2" or "xz". Default: FALSE.}
}
\description{
Create a table with key and val columns.
CREATE TABLE r_cache (
  key VARCHAR(128) PRIMARY KEY,
  val TEXT
)
}
\examples{

\dontrun{
pg <- cache_postgresql(pg_con, "r_cache")
mem_runif <- memoise(runif, cache = pg)
}


}
