% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_redis.R
\name{cache_redis}
\alias{cache_redis}
\title{Redis Cache}
\usage{
cache_redis(..., algo = "md5")
}
\arguments{
\item{...}{arguments passed along to the `hiredis()` function in the **redux** package.}

\item{algo}{The hashing algorithm used for the cache, see
\code{\link[digest]{digest}} for available algorithms.}
}
\description{
Use a Redis based cache to persist memoisation between R sessions.
}
\examples{

\dontrun{

my_function <- memoise::memoise(
  function(input) {
    Sys.sleep(5)
    paste('Input was', input)
  },
  cache = cache_redis()
)
my_function("a")
my_function("a")

}
}
\author{
Carson Sievert
}
