% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_datastore.r
\name{datastore_cache}
\alias{datastore_cache}
\title{datastore_cache}
\usage{
datastore_cache(cache_name = "rcache")
}
\description{
Use google datastore to store and retrieve cache items. Requires authentication.
}
\seealso{
\url{https://cloud.google.com/}

\url{https://cloud.google.com/datastore/docs/concepts/overview}

\url{https://developers.google.com/identity/protocols/OAuth2#basicsteps}
}

