% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/memoise.r
\name{has.cache}
\alias{has.cache}
\title{Test whether a memoised function has been cached for particular arguments.}
\usage{
has.cache(f, ...)
}
\arguments{
\item{f}{Function to test.}

\item{...}{arguments to function.}
}
\description{
Test whether a memoised function has been cached for particular arguments.
}
\examples{
mem_sum <- memoise(sum)
has.cache( mem_sum, 1, 2, 3) # FALSE
mem_sum(1, 2, 3)
has.cache( mem_sum, 1, 2, 3) # TRUE
}
\seealso{
\code{\link{is.memoised}}, \code{\link{memoise}}
}

