% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memoise.R
\name{has_cache}
\alias{has_cache}
\title{Test whether a memoised function has been cached for particular arguments.}
\usage{
has_cache(f)
}
\arguments{
\item{f}{Function to test.}
}
\value{
A function, with the same arguments as \code{f}, that can be called to test
  if \code{f} has cached results.
}
\description{
Test whether a memoised function has been cached for particular arguments.
}
\examples{
mem_sum <- memoise(sum)
has_cache(mem_sum)(1, 2, 3) # FALSE
mem_sum(1, 2, 3)
has_cache(mem_sum)(1, 2, 3) # TRUE
}
\seealso{
\code{\link{is.memoised}}, \code{\link{memoise}}
}
