% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memoise.R
\name{unmemoise}
\alias{unmemoise}
\title{convert memoised function back to unmemoised function}
\usage{
unmemoise(f)
}
\arguments{
\item{f}{the memoised function}
}
\value{
the unmemoised version of f
}
\description{
convert memoised function back to unmemoised function
}
\examples{
hello <- function() { print("Hello, world!") }
hello()
hello()
is.memoised(hello)
hello <- memoise(hello)
hello()
hello()
is.memoised(hello)
hello <- unmemoise(hello)
is.memoised(hello)
hello()
}
