% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_datastore.r
\name{cache_datastore}
\alias{cache_datastore}
\title{Google Datastore Cache}
\usage{
cache_datastore(project, cache_name = "rcache")
}
\arguments{
\item{project}{Google Cloud project}

\item{cache_name}{datastore kind to use for storing cache entities.}
}
\description{
Initiate a Google Datastore cache.
}
\examples{
\dontrun{
ds <- cache_datastore(project = "<project-id>", cache_name = "rcache")
mem_memoise(runif, cache = ds)
}

}
\seealso{
\url{https://cloud.google.com/}

\url{https://cloud.google.com/datastore/docs/concepts/overview}
}

