% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_datastore.r
\name{cache_datastore}
\alias{cache_datastore}
\title{Google Datastore Cache}
\usage{

}
\arguments{
\item{project}{Google Cloud project}

\item{cache_name}{datastore kind to use for storing cache entities.}
}
\description{
Initiate a Google Datastore cache.
}
\seealso{
\url{https://cloud.google.com/}

\url{https://cloud.google.com/datastore/docs/concepts/overview}
}

