% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_filesystem.r
\name{cache_filesystem}
\alias{cache_filesystem}
\title{Filesystem Cache}
\usage{
cache_filesystem(path)
}
\arguments{
\item{path}{Directory in which to store cached items.}
}
\description{
Initiate a filesystem cache.
}
\examples{

\dontrun{
# Use with Dropbox

db <- cache_filesystem("~/Dropbox/.rcache")

mem_runif <- memoise(runif, cache = db)

# Use with Google Drive

gd <- cache_filesystem("~/Google Drive/.rcache")

mem_runif <- memoise(runif, cache = gd)

}

}

