% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_filesystem.r
\name{cache_filesystem}
\alias{cache_filesystem}
\title{Filesystem Cache}
\usage{
cache_filesystem(path)
}
\arguments{
\item{path}{Directory in which to store cached items.}
}
\description{
Initiate a filesystem cache.
}

