% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_filesystem.r
\name{cache_filesystem}
\alias{cache_filesystem}
\title{Initiate a filesystem cache.}
\usage{
cache_filesystem stores cached items within a directory. It can be used with
Dropbox or Google Drive to allow for cache access from multiple computers.

To use cache_filesystem with Dropbox, you can specify:

  cache_filesystem("~/Dropbox/.rcache")

For Google Drive:

  cache_filesystem("~/Google Drive/.rcache")
}
\arguments{
\item{path}{Directory in which to store cached items.}
}

