% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_redis.r
\name{cache_redis}
\alias{cache_redis}
\title{Redis cache}
\usage{
cache_redis(..., index = 0)
}
\arguments{
\item{...}{Arguments passed to \code{rredis::redisConnect()}}

\item{index}{The nonnegative integer value of the redis database to select (default = 0).}
}
\description{
Initiate an redis cache
}
\examples{

\dontrun{
# Set up a unique bucket name.
r1 <- cache_redis()
mem_runif <- memoise(runif, cache = r1)
}


}

