% Generated by roxygen2 (4.0.0): do not edit by hand
\name{forget}
\alias{forget}
\title{Forget past results.
Resets the cache of a memoised function.}
\usage{
forget(f)
}
\arguments{
\item{f}{memoised function}
}
\description{
Forget past results.
Resets the cache of a memoised function.
}
\examples{
memX <- memoise(function() { Sys.sleep(1); runif(1) })
# The forget() function
system.time(print(memX()))
system.time(print(memX()))
forget(memX)
system.time(print(memX()))
}
\seealso{
\code{\link{memoise}}, \code{\link{is.memoised}}
}

