% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memoise.R
\name{forget}
\alias{forget}
\title{Forget past results.
Resets the cache of a memoised function.}
\usage{
forget(f)
}
\arguments{
\item{f}{memoised function}
}
\description{
NOTE: This resets the complete cache. If the cache was used for more than one function,
all cached results will be discarded.
}
\examples{
memX <- memoise(function() { Sys.sleep(1); runif(1) })
# The forget() function
system.time(print(memX()))
system.time(print(memX()))
forget(memX)
system.time(print(memX()))
}
\seealso{
\code{\link{memoise}}, \code{\link{is.memoised}}
}
